chore(webclient): refactors tests for components.
